tests/wpt/web-platform-tests/html/canvas/element/manual/wide-gamut-canvas/canvas-display-p3-drawImage-ImageBitmap-video.html

<!DOCTYPE HTML>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="canvas-display-p3.js"></script>
<body>
<script>
// Test that drawing ImageBitmaps with different video source color profiles
// into sRGB and Display P3 canvases works, by reading pixels with
// getImageData() as sRGB and Display P3 values.
for (let [filenameBase, expectedPixels] of Object.entries(videoTests)) {
    for (let contextColorSpace of ["srgb", "display-p3"]) {
        for (let imageDataColorSpace of ["srgb", "display-p3"]) {
            for (let cropSource of [false, true]) {
                async_test(function(t) {
                    let video = document.createElement("video");
                    video.onloadeddata = t.step_func(function() {

                        let canvas = document.createElement("canvas");
                        canvas.width = 2;
                        canvas.height = 2;

                        let ctx = canvas.getContext("2d", { colorSpace: contextColorSpace });

                        let imageBitmapPromise;
                        if (cropSource)
                            imageBitmapPromise = createImageBitmap(video, 1, 1, 1, 1);
                        else
                            imageBitmapPromise = createImageBitmap(video);

                        imageBitmapPromise.then(t.step_func_done(function(imageBitmap) {
                            video.remove();
                            ctx.drawImage(imageBitmap, 0, 0);

                            let imageData = ctx.getImageData(0, 0, 1, 1, { colorSpace: imageDataColorSpace });

                            let expected = expectedPixels[`${contextColorSpace} ${imageDataColorSpace}`];
                            assert_true(pixelsApproximatelyEqual(imageData.data, expected), `Actual pixel value ${[...imageData.data]} is approximately equal to ${expected}.`);
                        }), t.step_func(function(error) {
                            video.remove();
                            throw error;
                        }));
                    });
                    for (let format of ["mp4", "webm"]) {
                        let source = document.createElement("source");
                        source.src = `resources/${filenameBase}.${format}`;
                        source.type = `video/${format}`;
                        video.append(source);
                    }
                    document.body.append(video);
                }, `${filenameBase}, Context ${contextColorSpace}, ImageData ${imageDataColorSpace}, cropSource=${cropSource}`);
            }
        }
    }
}
</script>
