third_party/blink/web_tests/platform/win/external/wpt/html/canvas/element/manual/wide-gamut-canvas/canvas-display-p3-drawImage-ImageBitmap-video-expected.txt

This is a testharness.js-based test.
[FAIL] Rec2020-222000000, Context srgb, ImageData srgb, cropSource=false
  assert_true: Actual pixel value 169,0,0,255 is approximately equal to 186,0,0,255. expected true got false
[FAIL] Rec2020-222000000, Context srgb, ImageData srgb, cropSource=true
  assert_true: Actual pixel value 169,0,0,255 is approximately equal to 186,0,0,255. expected true got false
[FAIL] Rec2020-222000000, Context srgb, ImageData display-p3, cropSource=false
  assert_true: Actual pixel value 155,30,20,255 is approximately equal to 170,34,23,255. expected true got false
[FAIL] Rec2020-222000000, Context srgb, ImageData display-p3, cropSource=true
  assert_true: Actual pixel value 155,30,20,255 is approximately equal to 170,34,23,255. expected true got false
[FAIL] Rec2020-222000000, Context display-p3, ImageData srgb, cropSource=false
  assert_true: Actual pixel value 168,0,0,255 is approximately equal to 186,0,0,255. expected true got false
[FAIL] Rec2020-222000000, Context display-p3, ImageData srgb, cropSource=true
  assert_true: Actual pixel value 168,0,0,255 is approximately equal to 186,0,0,255. expected true got false
[FAIL] Rec2020-222000000, Context display-p3, ImageData display-p3, cropSource=false
  assert_true: Actual pixel value 153,0,4,255 is approximately equal to 169,0,3,255. expected true got false
[FAIL] Rec2020-222000000, Context display-p3, ImageData display-p3, cropSource=true
  assert_true: Actual pixel value 153,0,4,255 is approximately equal to 169,0,3,255. expected true got false
Harness: the test ran to completion.
